progresso evolução das habilidades

+++ src/entities/state/DeadState.java
@@ -3,7 +3,6 @@ package entities.state;
 import entities.ally.Ally;
 import entities.character.Player;
 import entities.skill.Skill;
-import entities.skill.attack.Trap;
 
 public class DeadState extends State {
 
@@ -24,7 +23,7 @@ public class DeadState extends State {
     }
 
     @Override
-    public double calculateDamage(Player activePlayer, Player passivePlayer, int activeSKillPowerAttack) {
+    public double calculateDamage(Player activePlayer, Player passivePlayer, double activeSKillPowerAttack) {
         return 0;
     }
